feat(fe): implement result badge for exercise detail
