+++ apps/frontend/app/(client)/(main)/course/[courseId]/_components/ExerciseAccordion.tsx
@@ -28,6 +28,7 @@ import Link from 'next/link'
 import { useEffect, useState } from 'react'
 import { useInterval } from 'react-use'
 import { AssignmentLink } from './AssignmentLink'
+import { ResultBadge } from './ResultBadge'
 
 interface ExerciseAccordionProps {
   courseId: number
@@ -298,38 +299,6 @@ function SubmissionBadge({ className, grade }: SubmissionBadgeProps) {
           {grade.submittedCount}/{grade.problemCount}
         </p>
       </div>
-    </div>
-  )
-}
-
-interface ResultBadgeProps {
-  className?: string
-  assignmentSubmission: AssignmentSubmission
-}
-
-function ResultBadge({ className, assignmentSubmission }: ResultBadgeProps) {
-  const submissionResult =
-    assignmentSubmission?.submission?.submissionResult ?? 'No Result'
-
-  let badgeStyle = ''
-
-  if (submissionResult === 'Accepted') {
-    badgeStyle = 'border border-primary text-primary'
-  } else if (submissionResult === 'No Result') {
-    badgeStyle = 'border border-[#B0B0B0] text-[#B0B0B0]'
-  } else {
-    badgeStyle = 'border border-[#FC5555] text-[#FC5555]'
-  }
-
-  return (
-    <div
-      className={cn(
-        'flex h-[38px] w-[140px] items-center justify-center rounded-full border',
-        badgeStyle,
-        className
-      )}
-    >
-      <p className="text-sm font-medium">{submissionResult}</p>
     </div>
   )
 }
